Bug 1090894 - Workaround for the charger led bug
This works around the charger led bug which appeared with revision B03
of ZTE's firmware for the Open C.

It directly puts the leds gpios under control by the CPU instead of
delegating this task the the modem firmware.

However, this fix still isn't quite perfect as the CPU cannot blink leds
while suspended, and there is no reasonable way to work around that.